Re-organize files under `ansible/devutil` to have better structure for future enhancements (#8917)
What is the motivation for this PR?
Re-organize the files under ansible/devutil to have a clearer structure for future enhancements.

How did you do it?
Moved and renamed files.

How did you verify/test it?
Test run the ansible/upgrade_sonic.py and .azure-pipeline/upgrade_image.py scripts.

Any platform specific information?
NA
