+++ ansible/devutil/devices/sonic.py
@@ -1,12 +1,29 @@
 import logging
-import os
+import yaml
 
-from devutil.ansible_hosts import RunAnsibleModuleFailed
+from .ansible_hosts import AnsibleHosts
+from .ansible_hosts import RunAnsibleModuleFailed
 
 logger = logging.getLogger(__name__)
 
-_self_dir = os.path.dirname(os.path.abspath(__file__))
-ansible_path = os.path.realpath(os.path.join(_self_dir, "../"))
+
+class SonicHosts(AnsibleHosts):
+    SUPPORTED_UPGRADE_TYPES = ["onie", "sonic"]
+
+    def __init__(self, inventories, host_pattern, options={}, hostvars={}):
+        super(SonicHosts, self).__init__(inventories, host_pattern, options=options.copy(), hostvars=hostvars.copy())
+
+    @property
+    def sonic_version(self):
+        try:
+            output = self.command("cat /etc/sonic/sonic_version.yml")
+            versions = {}
+            for hostname in self.hostnames:
+                versions[hostname] = yaml.safe_load(output[hostname]["stdout"])
+            return versions
+        except Exception as e:
+            logger.error("Failed to run `cat /etc/sonic/sonic_version.yml`: {}".format(repr(e)))
+            return {}
 
 
 def upgrade_by_sonic(sonichosts, image_url, disk_used_percent):
